+++ pkg/api/plugins/template/apis/template/register.go
@@ -1,0 +1,6 @@
+package team
+
+// GroupName is the group name used in this package.
+const (
+	GroupName = "kobs.io"
+)